refactor: rename types.ts to items.ts
