+++ packages/api-postgres/src/common/types/engagementItems.ts
@@ -1,5 +1,5 @@
 import { ItemCategory } from './enums';
-import { BaseItemFields } from './types';
+import { BaseItemFields } from './items';
 
 // Engagement Items
 export const availableEngagementItems = ['games', 'artwork', 'candy/treats', 'other'] as const;
