+++ packages/api-postgres/src/common/types/personalHygieneItems.ts
@@ -1,5 +1,5 @@
 import { ItemCategory } from './enums';
-import { BaseItemFields } from './types';
+import { BaseItemFields } from './items';
 
 // Personal Hygiene Items
 export const availablePersonalHygieneItems = [
