+++ packages/api-postgres/src/common/types/petItems.ts
@@ -1,5 +1,5 @@
 import { ItemCategory } from './enums';
-import { BaseItemFields } from './types';
+import { BaseItemFields } from './items';
 
 // Pet Items
 export const availablePetItems = ['other'] as const;
